fix(kno-9922): move custom query serialization to src/lib so it survives regeneration
The KNO-9922 split serialization (scalar arrays -> brackets, object arrays ->
indices) lived directly in the generated `src/internal/utils/query.ts`. Stainless
regenerates that file, and commit be1705d ("chore(formatter): run prettier and
eslint separately") on the 1.33.0 release branch overwrote it back to plain
`arrayFormat: 'brackets'`, reintroducing `objects[][id]` and failing
tests/stringifyQuery.test.ts.

Per CONTRIBUTING.md, only `src/lib/` and `examples/` are never touched by the
generator. Move the actual logic there (src/lib/stringifyQuery.ts) and have the
generated util re-export it, so the substantive code is in the generator-safe
zone. The one-line re-export in the generated util is the only remaining
generated-file touchpoint; if a future regeneration drops it,
tests/stringifyQuery.test.ts fails loudly and it's a one-line re-add.

Note: fully durable protection still requires restoring this customization in the
Stainless config so the generator emits the re-export on every run.
